requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
cloudinary
torch
clip-by-openai
Pillow
numpy


